Moved worker to separate file
